+++ test/active_storage_test.rb
@@ -54,6 +54,8 @@ class ActiveStorageTest < Minitest::Test
   end
 
   def test_encrypt_create
+    skip if ActiveStorage::VERSION::MAJOR >= 6
+
     message = "hello world"
 
     file = Tempfile.new